refactor(frontend): extract SystemPanel from page.tsx (P1.d)
Largest extract to date. Moves the /system route UI — systemConfig
state + ~30 editable fields, workerMonitor state + 5s poll, client-
connection form (apiBase/tenantId/accessToken) — into
app/components/SystemPanel.tsx. page.tsx drops 890 lines net.

Behavior preserved, coupling reduced:

- SystemPanel owns systemConfig, workerMonitor, isWorkerMonitorLoading,
  apiBase, tenantId, accessToken. All allocated only while /system is
  active; navigating away unmounts and terminates the worker-monitor
  poll interval.
- Global setStatusMessage/setErrorMessage flow in as onStatus/onError.
- systemStatus (used for global banner warnings even off /system) still
  lives on HomePage. SystemPanel reads it via prop for the header pills
  and triggers a parent refresh via onSystemStatusRefresh when config
  or secret writes land (which can flip redis/llm health).
- handleTenantSave calls back to onTenantConnectionSaved so the parent
  refreshes its data after localStorage identity changes — replaces the
  prior \`[tenantId]\`-dep effect, which is now impossible since tenantId
  is panel-local.
- buildConfigPayload dedupes the 30-field payload that previously lived
  in two near-identical copies (handleSystemConfigSave and
  handleSaveSecret).

Verified: frontend 46/46 (stable across 5 full-suite reruns),
production build succeeds, /system route renders and saves config
identically in smoke test.
